chore(decider,tests): store sync info [NET-639] (#130)
* store synced info
* refactor consts
* refactor tests
* refactor worker settings
