+++ src/aqua/decider/consts.aqua
@@ -14,3 +14,12 @@ const FAILED_DEALS = "failed_deals"
 -- List of joined deals. For each deal from the list a worker was created and
 -- subnet registration request was sent
 const JOINED_DEALS = "joined_deals"
+
+-- The block from chain decider processed.
+const LAST_SEEN_BLOCK = "last_seen_block"
+
+-- Decider puts in Worker Spell's KV storage CID of worker definition
+-- This name must be in sync with Worker Spell.
+const WORKER_DEF_CID = "worker_def_cid"
+
+const SYNC_INFO = "sync_info"
