+++ src/aqua/decider/deal_storage.aqua
@@ -17,7 +17,6 @@ service JsonDealState("json"):
 
 func store_deal_state(decider_id: SpellId, deal_id: DealId, state: DealState):
     Spell decider_id
-    deal_log(decider_id, deal_id, ["store left boundary", state.left_boundary])
 
     json <- Json.stringify(state)
     Spell.set_string(deal_id, json)
