+++ src/tests/decider-distro-tests-rs/tests/test_deploy.rs
@@ -11,9 +11,17 @@ use serde::Deserialize;
 use serde_json::{json, Value};
 use std::collections::HashSet;
 use std::time::Duration;
+use utils::chain::LogsReq;
+use utils::control::{
+    update_config, update_decider_script_for_tests, wait_decider_stopped, wait_worker_spell_stopped,
+};
+use utils::deal::{get_deal_state, get_joined_deals, JoinedDeal};
+use utils::default::{default_receipt, DEAL_IDS, DEFAULT_POLL_WINDOW_BLOCK_SIZE};
+use utils::distro::{make_distro_with_api, make_distro_with_api_and_config};
+use utils::setup::setup_nox;
 use utils::test_rpc_server::{run_test_server, run_test_server_predefined};
+use utils::worker::get_worker_app_cid;
 use utils::TestApp;
-use utils::DEAL_IDS;
 use utils::*;
 
 /// Test the basic flow
@@ -56,6 +64,7 @@ use utils::*;
 ///
 #[tokio::test]
 async fn test_deploy_a_deal_single() {
+    enable_decider_logs();
     const DEAL_ID: &'static str = DEAL_IDS[0];
     const BLOCK: u32 = 32;
     const LATEST_BLOCK: u32 = 35;
@@ -97,12 +106,9 @@ async fn test_deploy_a_deal_single() {
     let mut result = execute(
         &mut client,
         r#"
-            (seq
                 (call relay ("decider" "get_u32") ["counter"] counter)
-                (call relay ("decider" "list_get_strings") ["joined_deals"] deals)
-            )
         "#,
-        "counter deals",
+        "counter",
         hashmap! {},
     )
     .await
@@ -118,7 +124,7 @@ async fn test_deploy_a_deal_single() {
 
     // Analyse joined deals
     let deal = {
-        let mut deals = parse_joined_deals(result.remove(0));
+        let mut deals = get_joined_deals(&mut client).await;
         assert_eq!(deals.len(), 1, "decider joined more than one deal");
         deals.remove(0)
     };
