+++ src/tests/decider-distro-tests-rs/tests/tests_basic.rs
@@ -6,15 +6,23 @@ pub mod utils;
 
 use utils::test_rpc_server::run_test_server;
 
-use crate::utils::*;
+use crate::utils::default::DEFAULT_POLL_WINDOW_BLOCK_SIZE;
 use connected_client::ConnectedClient;
 use created_swarm::make_swarms_with_cfg;
+use eyre::WrapErr;
 use fluence_app_service::TomlMarineConfig;
 use fluence_spell_dtos::trigger_config::TriggerConfig;
-use fluence_spell_dtos::value::{StringValue, UnitValue};
+use fluence_spell_dtos::value::UnitValue;
 use maplit::hashmap;
 use serde::Serialize;
 use serde_json::{json, Value};
+use utils::chain::LogsReq;
+use utils::control::{update_config, update_decider_script_for_tests, wait_decider_stopped};
+use utils::deal::get_joined_deals;
+use utils::default::{default_receipt, DEAL_IDS};
+use utils::distro::*;
+use utils::setup::setup_nox;
+use utils::*;
 
 #[test]
 fn test_connector_config_check() {
@@ -235,15 +243,12 @@ async fn test_left_boundary_idle() {
     // To be able to wait 'til the end of one cycle
     update_decider_script_for_tests(&mut client, swarm.tmp_dir.clone()).await;
 
-    let mut oneshot_config = TriggerConfig::default();
-    oneshot_config.clock.start_sec = 1;
-
     let block_numbers = vec!["0x0", "0x10", "0x10", "0xffffff"];
     let expected_last_seen = vec!["0x0", "0x10", "0x10", "0x205"];
     let expected_from_blocks = vec!["0x0", "0x1", "0x11", "0x11"];
 
     for step in 0..block_numbers.len() {
-        update_config(&mut client, &oneshot_config).await.unwrap();
+        update_config(&mut client, &oneshot_config()).await.unwrap();
         {
             let (method, params) = server.receive_request().await.unwrap();
             assert_eq!(method, "eth_blockNumber");
@@ -262,17 +267,84 @@ async fn test_left_boundary_idle() {
         }
         wait_decider_stopped(&mut client).await;
 
-        let result = execute(
-            &mut client,
-            r#" (call relay ("decider" "get_string") ["last_seen_block"] last_seen) "#,
-            "last_seen",
-            hashmap! {},
-        )
-        .await
-        .unwrap();
-        let last_seen = serde_json::from_value::<StringValue>(result[0].clone()).unwrap();
+        let last_seen = spell::get_string(&mut client, "decider", "last_seen_block")
+            .await
+            .unwrap();
         assert_eq!(last_seen.str, expected_last_seen[step]);
     }
+
+    server.shutdown().await;
+}
+
+#[tokio::test]
+async fn test_sync_info() {
+    const LATEST_BLOCK_FIRST_RUN: u32 = 100;
+    const LATEST_BLOCK_SECOND_RUN: u32 = 2000;
+
+    let mut server = run_test_server();
+
+    let url = server.url.clone();
+
+    let distro = make_distro_with_api(url);
+    let (swarm, mut client) = setup_nox(distro.clone()).await;
+
+    update_decider_script_for_tests(&mut client, swarm.tmp_dir.clone()).await;
+    update_config(&mut client, &oneshot_config()).await.unwrap();
+    {
+        {
+            let (method, _params) = server.receive_request().await.unwrap();
+            assert_eq!(method, "eth_blockNumber");
+            server.send_response(Ok(json!(to_hex(LATEST_BLOCK_FIRST_RUN))));
+        }
+        {
+            let (method, _params) = server.receive_request().await.unwrap();
+            assert_eq!(method, "eth_getLogs");
+            server.send_response(Ok(json!([])));
+        }
+    }
+    wait_decider_stopped(&mut client).await;
+
+    let sync_info = decider::get_sync_info(&mut client)
+        .await
+        .wrap_err("get_sync_info")
+        .unwrap();
+    assert_eq!(
+        sync_info.run_updated, 1,
+        "should be updated on the first run"
+    );
+    assert_eq!(sync_info.blocks_diff, 0, "must be in sync");
+
+    update_config(&mut client, &oneshot_config()).await.unwrap();
+    {
+        {
+            let (method, _params) = server.receive_request().await.unwrap();
+            assert_eq!(method, "eth_blockNumber");
+            server.send_response(Ok(json!(to_hex(LATEST_BLOCK_SECOND_RUN))));
+        }
+        {
+            let (method, _params) = server.receive_request().await.unwrap();
+            assert_eq!(method, "eth_getLogs");
+            server.send_response(Ok(json!([])));
+        }
+    }
+    wait_decider_stopped(&mut client).await;
+
+    let sync_info = decider::get_sync_info(&mut client)
+        .await
+        .wrap_err("get_sync_info")
+        .unwrap();
+    assert_eq!(
+        sync_info.run_updated, 2,
+        "should be updated on the first run"
+    );
+
+    let expected_last_seen = LATEST_BLOCK_FIRST_RUN + DEFAULT_POLL_WINDOW_BLOCK_SIZE;
+    assert_eq!(
+        sync_info.blocks_diff,
+        LATEST_BLOCK_SECOND_RUN - expected_last_seen - 1,
+        "must be not in sync with {} block range",
+        DEFAULT_POLL_WINDOW_BLOCK_SIZE
+    );
 
     server.shutdown().await;
 }
