+++ src/tests/decider-distro-tests-rs/tests/utils/mod.rs
@@ -1,85 +1,27 @@
+pub mod control;
+pub mod deal;
+pub mod distro;
+pub mod setup;
+pub mod subnet;
+
+pub mod chain;
+
+pub mod default;
+pub mod worker;
+
+pub mod spell;
+
+pub mod decider;
+
 mod test_apps;
 pub mod test_rpc_server;
 
 pub use test_apps::TestApp;
 
 use connected_client::ConnectedClient;
-use created_swarm::system_services_config::{AquaIpfsConfig, SystemServicesConfig};
-use created_swarm::{make_swarms_with_cfg, CreatedSwarm};
-use decider_distro::DeciderConfig;
-use eyre::WrapErr;
-use fluence_app_service::TomlMarineConfig;
 use fluence_spell_dtos::trigger_config::TriggerConfig;
-use fluence_spell_dtos::value::{ScriptValue, StringListValue, StringValue};
-use maplit::hashmap;
-use serde::Deserialize;
 use serde_json::{json, Value};
 use std::collections::HashMap;
-use std::fs;
-use std::path::PathBuf;
-use std::time::Duration;
-use system_services::{PackageDistro, ServiceDistro, SpellDistro};
-
-pub const DEFAULT_POLL_WINDOW_BLOCK_SIZE: u32 = 500;
-
-pub const DEAL_IDS: &[&'static str] = &[
-    "ffa0611a099ab68ad7c3c67b4ca5bbbee7a58b99",
-    "880a53a54785df22ba804aee81ce8bd0d45bdedc",
-    "67b2ad3866429282e16e55b715d12a77f85b7ce8",
-    "1234563866429282e16e55b715d12a77f85b7cc9",
-    "991b64a54785df22ba804aee81ce8bd0d45bdabb",
-    "3665748409e712cd91b428c18e07a8e37b44c47e",
-];
-
-pub const IPFS_MULTIADDR: &str = "/ip4/127.0.0.1/tcp/5001";
-
-pub fn default_receipt() -> Value {
-    json!({"status" : "0x1", "blockNumber": "0x300"})
-}
-
-pub fn setup_aqua_ipfs() -> AquaIpfsConfig {
-    let mut config = AquaIpfsConfig::default();
-    static IPFS_CLI_PATH: Option<&str> = option_env!("IPFS_CLI_PATH");
-    if let Some(path) = IPFS_CLI_PATH {
-        config.ipfs_binary_path = path.to_string();
-    }
-    config.external_api_multiaddr = IPFS_MULTIADDR.to_string();
-    config.local_api_multiaddr = IPFS_MULTIADDR.to_string();
-    config
-}
-
-pub fn setup_system_config() -> SystemServicesConfig {
-    let mut config = SystemServicesConfig::default();
-    config.aqua_ipfs = setup_aqua_ipfs();
-    config
-}
-
-pub async fn setup_swarm(distro: PackageDistro) -> CreatedSwarm {
-    let mut swarms = make_swarms_with_cfg(1, move |mut cfg| {
-        cfg.enabled_system_services = vec!["aqua-ipfs".to_string()];
-        cfg.extend_system_services = vec![distro.clone()];
-        let config = setup_system_config();
-        cfg.allowed_binaries = vec![
-            config.aqua_ipfs.ipfs_binary_path.clone(),
-            config.connector.curl_binary_path.clone(),
-        ];
-        cfg.override_system_services_config = Some(config);
-        cfg
-    })
-    .await;
-    swarms.remove(0)
-}
-
-pub async fn setup_nox(distro: PackageDistro) -> (CreatedSwarm, ConnectedClient) {
-    let swarm = setup_swarm(distro).await;
-    let client = ConnectedClient::connect_with_keypair(
-        swarm.multiaddr.clone(),
-        Some(swarm.management_keypair.clone()),
-    )
-    .await
-    .unwrap();
-    (swarm, client)
-}
 
 pub fn enable_decider_logs() {
     let namespaces = vec![
@@ -109,94 +51,6 @@ pub fn enable_decider_logs() {
     log_utils::enable_logs_for(spec);
 }
 
-pub fn package_items_names(distro: &PackageDistro) -> Vec<String> {
-    distro
-        .services
-        .iter()
-        .map(|s| s.name.clone())
-        .chain(distro.spells.iter().map(|s| s.name.clone()))
-        .collect()
-}
-
-pub fn make_distro(trigger_config: TriggerConfig, settings: DeciderConfig) -> PackageDistro {
-    let connector = decider_distro::connector_service_modules();
-    let marine_config: TomlMarineConfig =
-        toml::from_slice(connector.config).expect("parse marine config");
-    let service = ServiceDistro {
-        modules: connector.modules,
-        config: marine_config,
-        name: connector.name.to_string(),
-    };
-
-    let distro_spell = decider_distro::decider_spell(settings);
-    let spell = SpellDistro {
-        name: "decider".to_string(),
-        air: distro_spell.air,
-        kv: distro_spell.kv.clone(),
-        trigger_config,
-    };
-
-    PackageDistro {
-        name: "decider".to_string(),
-        version: decider_distro::VERSION,
-        services: vec![service],
-        spells: vec![spell],
-        init: None,
-    }
-}
-
-pub fn make_distro_default() -> PackageDistro {
-    let decider_settings = DeciderConfig {
-        worker_period_sec: 0,
-        worker_ipfs_multiaddr: IPFS_MULTIADDR.to_string(),
-        chain_api_endpoint: "http://127.0.0.1:12009".to_string(),
-        chain_network_id: 11,
-        chain_contract_block_hex: "0x0".to_string(),
-        chain_matcher_addr: "0x0".to_string(),
-        chain_workers_gas: 210_00,
-        chain_wallet_key: "0x0".to_string(),
-    };
-    // let's try to run a decider cycle on demand by updating the config
-    let mut trigger_config = TriggerConfig::default();
-    trigger_config.clock.start_sec = 1;
-    make_distro(trigger_config, decider_settings)
-}
-
-pub fn make_distro_with_api(api: String) -> PackageDistro {
-    let decider_settings = DeciderConfig {
-        // worker will run once
-        worker_period_sec: 0,
-        worker_ipfs_multiaddr: IPFS_MULTIADDR.to_string(),
-        chain_api_endpoint: api,
-        chain_network_id: 11,
-        chain_contract_block_hex: "0x0".to_string(),
-        chain_matcher_addr: "0x0".to_string(),
-        chain_workers_gas: 210_00,
-        chain_wallet_key: "0xfdc4ba94809c7930fe4676b7d845cbf8fa5c1beae8744d959530e5073004cf3f"
-            .to_string(),
-    };
-    // decider will run once
-    let trigger_config = TriggerConfig::default();
-    make_distro(trigger_config, decider_settings)
-}
-
-pub fn make_distro_with_api_and_config(api: String, config: TriggerConfig) -> PackageDistro {
-    let decider_settings = DeciderConfig {
-        // worker will run once
-        worker_period_sec: 0,
-        worker_ipfs_multiaddr: IPFS_MULTIADDR.to_string(),
-        chain_api_endpoint: api,
-        chain_network_id: 11,
-        chain_contract_block_hex: "0x0".to_string(),
-        chain_matcher_addr: "0x0".to_string(),
-        chain_workers_gas: 210_00,
-        chain_wallet_key: "0xfdc4ba94809c7930fe4676b7d845cbf8fa5c1beae8744d959530e5073004cf3f"
-            .to_string(),
-    };
-    // decider will run once
-    make_distro(config, decider_settings)
-}
-
 pub async fn execute(
     client: &mut ConnectedClient,
     correct_air: &str,
@@ -214,374 +68,10 @@ pub async fn execute(
         .await
 }
 
-pub async fn update_config(
-    client: &mut ConnectedClient,
-    trigger_config: &TriggerConfig,
-) -> eyre::Result<Vec<Value>> {
-    execute(
-        client,
-        r#"(call relay ("spell" "update_trigger_config") ["decider" config])"#,
-        "\"done\"",
-        hashmap! {
-            "config" => json!(trigger_config),
-        },
-    )
-    .await
-}
-
-// God left me here
-pub fn modify_decider_spell_script(
-    tmp_dir: PathBuf,
-    decider_spell_id: String,
-    updated_script: String,
-) {
-    let script_path: PathBuf = tmp_dir.join(
-        [
-            "services",
-            "workdir",
-            &decider_spell_id,
-            "tmp",
-            "script.air",
-        ]
-        .iter()
-        .collect::<PathBuf>(),
-    );
-
-    fs::write(&script_path, updated_script).unwrap();
-}
-
-pub async fn update_decider_script_for_tests(client: &mut ConnectedClient, test_dir: PathBuf) {
-    let result = execute(
-        client,
-        r#"
-            (seq
-                (call relay ("srv" "resolve_alias_opt") ["decider"] id)
-                (call relay ("decider" "get_script_source_from_file") [] script)
-            )
-    "#,
-        "id script",
-        hashmap! {},
-    )
-    .await
-    .unwrap();
-    assert_eq!(
-        result[0].as_array().unwrap().len(),
-        1,
-        "can't resolve `decider` alias"
-    );
-    let decider_id = result[0].as_array().unwrap()[0]
-        .as_str()
-        .unwrap()
-        .to_string();
-    let script = serde_json::from_value::<ScriptValue>(result[1].clone()).unwrap();
-    assert!(script.success, "can't get decider script");
-
-    let updated_script = format!(
-        r#"
-        (seq
-            {script}
-            (call "{client}" ("return" "") ["done"])
-        )
-    "#,
-        client = client.peer_id,
-        script = script.source_code,
-    );
-
-    modify_decider_spell_script(test_dir, decider_id, updated_script);
-}
-
 pub fn oneshot_config() -> TriggerConfig {
     let mut config = TriggerConfig::default();
     config.clock.start_sec = 1;
     config
-}
-
-pub async fn wait_worker_spell_stopped(
-    client: &mut ConnectedClient,
-    worker_id: String,
-    timeout_per_try: Duration,
-) {
-    let mut finished = false;
-    for _ in 0..5 {
-        // if only we can import these keys from Aqua files
-        let result = execute(
-            client,
-            r#"
-            (seq
-                (call relay ("op" "noop") [])
-                (call worker ("worker-spell" "list_get_strings") ["__installation_spell_status__"] status)
-            )
-        "#,
-            r#"status"#,
-            hashmap! {
-                "worker" => json!(worker_id),
-            },
-        )
-            .await
-            .wrap_err("getting installation spell status")
-            .unwrap();
-
-        assert!(!result.is_empty(), "no result from the worker-spell");
-
-        let strings = serde_json::from_value::<StringListValue>(result[0].clone()).unwrap();
-        assert!(
-            strings.success,
-            "can't get installation spell status: {}",
-            strings.error
-        );
-
-        if !strings.strings.is_empty() {
-            #[derive(Deserialize, Debug)]
-            struct State {
-                state: String,
-            }
-            let last_status = strings.strings.last().unwrap();
-            let state = serde_json::from_str::<State>(last_status).unwrap();
-            let in_progress_statuses = ["INSTALLATION_IN_PROGRESS", "NOT_STARTED"];
-            if !in_progress_statuses.contains(&state.state.as_str()) {
-                assert_eq!(
-                    state.state, "INSTALLATION_SUCCESSFUL",
-                    "wrong installation spell status"
-                );
-                finished = true;
-                break;
-            }
-        }
-        tokio::time::sleep(timeout_per_try).await;
-    }
-    assert!(
-        finished,
-        "installation spell didn't finish in time or failed"
-    );
-}
-
-pub async fn wait_decider_stopped(client: &mut ConnectedClient) {
-    let decider_stopped = client
-        .receive_args()
-        .await
-        .wrap_err("wait decider")
-        .unwrap();
-
-    assert_eq!(
-        decider_stopped[0].as_str().unwrap(),
-        "done",
-        "decider ended with a different return status"
-    );
-}
-
-pub async fn get_worker_app_cid(client: &mut ConnectedClient, worker_id: &String) -> String {
-    let mut result = execute(
-        client,
-        r#"
-        (seq
-            (call relay ("op" "noop") [])
-            (call worker_id ("worker-spell" "get_string") ["worker_def_cid"] cid)
-        )
-        "#,
-        "cid",
-        hashmap! {
-            "worker_id" => json!(worker_id),
-        },
-    )
-    .await
-    .unwrap();
-    let result = serde_json::from_value::<StringValue>(result.remove(0)).unwrap();
-    assert!(!result.absent, "worker-spell doesn't have worker_def_cid");
-    serde_json::from_str::<String>(&result.str).unwrap()
-}
-
-#[derive(Deserialize, Debug)]
-pub struct DealState {
-    pub left_boundary: String,
-}
-
-pub async fn get_deal_state(client: &mut ConnectedClient, deal_id: &String) -> DealState {
-    let mut result = execute(
-        client,
-        r#"
-            (call relay ("decider" "get_string") [deal_id] deal_state)
-        "#,
-        "deal_state",
-        hashmap! {
-            "deal_id" => json!(deal_id)
-        },
-    )
-    .await
-    .unwrap();
-    let str = serde_json::from_value::<StringValue>(result.remove(0))
-        .wrap_err("get deal_state")
-        .unwrap();
-    assert!(!str.absent, "no state for deal {}", deal_id);
-    assert!(
-        str.success,
-        "can't get state for deal {}: {}",
-        deal_id, str.error
-    );
-    serde_json::from_str::<DealState>(&str.str)
-        .wrap_err("parse deal_state")
-        .unwrap()
-}
-
-#[derive(Deserialize, Debug)]
-pub struct JoinedDeal {
-    pub deal_id: String,
-    pub worker_id: String,
-}
-
-pub fn parse_joined_deals(deals: Value) -> Vec<JoinedDeal> {
-    let deals = serde_json::from_value::<StringListValue>(deals).unwrap();
-    assert!(deals.success);
-    deals
-        .strings
-        .iter()
-        .map(|deal| serde_json::from_str::<JoinedDeal>(deal).unwrap())
-        .collect()
-}
-
-pub async fn get_joined_deals(mut client: &mut ConnectedClient) -> Vec<JoinedDeal> {
-    let mut deals = execute(
-        &mut client,
-        r#"
-            (call relay ("decider" "list_get_strings") ["joined_deals"] deals)
-        "#,
-        "deals",
-        hashmap! {},
-    )
-    .await
-    .unwrap();
-    parse_joined_deals(deals.remove(0))
-}
-
-// atm the we don't use some fields in the tests, but will do in future
-#[allow(dead_code)]
-#[derive(Deserialize, Debug)]
-pub struct WorkerTxInfo {
-    deal_id: String,
-    tx_hash: String,
-}
-
-pub async fn get_txs(mut client: &mut ConnectedClient) -> Vec<WorkerTxInfo> {
-    let mut result = execute(
-        &mut client,
-        r#"
-            (call relay ("decider" "list_get_strings") ["worker_registration_txs"] txs)
-        "#,
-        "txs",
-        hashmap! {},
-    )
-    .await
-    .unwrap();
-    let txs = serde_json::from_value::<StringListValue>(result.remove(0)).unwrap();
-    assert!(
-        txs.success,
-        "can't receive `worker_registration_txs`: {}",
-        txs.error
-    );
-    txs.strings
-        .iter()
-        .map(|tx| serde_json::from_str::<WorkerTxInfo>(tx).unwrap())
-        .collect::<Vec<_>>()
-}
-
-#[allow(dead_code)]
-#[derive(Deserialize, Debug)]
-pub struct WorkerTxStatus {
-    tx_info: WorkerTxInfo,
-    status: String,
-}
-
-pub async fn get_txs_statuses(mut client: &mut ConnectedClient) -> Vec<WorkerTxStatus> {
-    let mut result = execute(
-        &mut client,
-        r#"
-            (call relay ("decider" "list_get_strings") ["worker_registration_txs_statuses"] txs)
-        "#,
-        "txs",
-        hashmap! {},
-    )
-    .await
-    .unwrap();
-    let txs = serde_json::from_value::<StringListValue>(result.remove(0)).unwrap();
-    assert!(
-        txs.success,
-        "can't receive `worker_registration_txs_statuses`: {}",
-        txs.error
-    );
-    txs.strings
-        .iter()
-        .map(|tx| serde_json::from_str::<WorkerTxStatus>(tx).unwrap())
-        .collect::<Vec<_>>()
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(tag = "type", content = "content")]
-pub enum FailedDealPayload {
-    InstallationFailed { log: Value },
-    TxFailed { tx_hash: Vec<String> },
-}
-
-#[allow(dead_code)]
-#[derive(Deserialize, Debug)]
-pub struct FailedDeal {
-    deal_id: String,
-    message: String,
-    payload: FailedDealPayload,
-}
-
-pub async fn get_failed_deals(mut client: &mut ConnectedClient) -> Vec<FailedDeal> {
-    let mut deals = execute(
-        &mut client,
-        r#"
-            (call relay ("decider" "list_get_strings") ["failed_deals"] deals)
-        "#,
-        "deals",
-        hashmap! {},
-    )
-    .await
-    .unwrap();
-    let lst = serde_json::from_value::<StringListValue>(deals.remove(0)).unwrap();
-    assert!(lst.success, "can't receive failed_deals: {}", lst.error);
-    lst.strings
-        .iter()
-        .map(|s| serde_json::from_str::<FailedDeal>(s))
-        .collect::<Result<Vec<_>, _>>()
-        .unwrap()
-}
-
-pub async fn get_worker(mut client: &mut ConnectedClient, deal: &str) -> Vec<String> {
-    let mut worker = execute(
-        &mut client,
-        r#"
-            (call relay ("worker" "get_worker_id") [dealid] worker)
-        "#,
-        "worker",
-        hashmap! {
-            "dealid" => json!(format!("0x{deal}"))
-        },
-    )
-    .await
-    .unwrap();
-    serde_json::from_value::<Vec<String>>(worker.remove(0)).unwrap()
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct LogsReq {
-    pub address: String,
-    #[serde(deserialize_with = "hex_u32_deserialize")]
-    pub from_block: u32,
-    #[serde(deserialize_with = "hex_u32_deserialize")]
-    pub to_block: u32,
-    pub topics: Vec<String>,
-}
-
-pub fn filter_logs<'a, T>(blocks: &'a [(u32, T)], req: &LogsReq) -> Vec<&'a (u32, T)> {
-    blocks
-        .iter()
-        .filter(|(block_number, _)| {
-            *block_number >= req.from_block && *block_number <= req.to_block
-        })
-        .collect::<_>()
 }
 
 pub fn hex_u32_deserialize<'de, D>(deserializer: D) -> Result<u32, D::Error>
